catalog/requirements.txt
uvicorn
pydantic
fastapi
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
requests
kafka-python==2.0.2
aiokafka==0.8.1
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-propagator-b3==1.21.0
httpx==0.25.2
pydantic-settings==2.1.0
pyjwt==2.10.1
elasticsearch[async]>=8.11.1,<9.0.0
aiohttp>=3.8.0
